TextToVideo/backend/requirements.txt
# Azure Services
azure-cognitiveservices-vision-computervision==0.9.0
azure-storage-blob==12.16.0
azure-identity==1.12.0
azure-functions==1.15.0
azure-media-videoanalyzer-edge==1.0.0b1
azure-core==1.32.0

# Video Generation
runwayml==3.4.0

# Web Framework
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1

# AI/ML
torch==2.1.0
torchvision==0.16.0
opencv-python==4.9.0.80
numpy>=1.26.4,<2.0.0

# Utilities
pydantic>=2.7.4,<3.0.0
langchain==0.3.19
langchain-core==0.3.40

# ML and Vision
ffmpeg-python==0.2.0
transformers==4.38.2
pillow==10.0.1
moviepy==1.0.3
requests==2.31.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1 